+++ README.md
@@ -1,26 +1,26 @@
-# DeepBLAST 
+# TM-Vec
 
 Learning protein structural similarity from sequence alone.  Our preprint can be found [here](https://www.biorxiv.org/content/10.1101/2020.11.03.365932v1)
 
 # Installation
 
-DeepBLAST can be installed from pip via
+TM-Vec can be installed from pip via
 
 ```
-pip install deepblast
+pip install tmvec
 ```
 
 To install from the development branch run
 
 ```
-pip install git+https://github.com/flatironinstitute/deepblast.git
+pip install git+https://github.com/flatironinstitute/tmvec.git
 ```
 
 # Downloading pretrained models and data
 
-The pretrained DeepBLAST model can be downloaded [here](https://users.flatironinstitute.org/jmorton/public_www/deepblast-public-data/checkpoints/deepblast-lstm4x.pt).
+The pretrained TM-vec model can be downloaded [here](https://users.flatironinstitute.org/jmorton/public_www/deepblast-public-data/checkpoints/deepblast-lstm4x.pt).
 
-The TM-align structural alignments used to pretrain DeepBLAST can be found [here](https://users.flatironinstitute.org/jmorton/public_www/deepblast-public-data/tmalign.tar.gz)
+The TM-align structural alignments used to pretrain Tmvec can be found [here](https://users.flatironinstitute.org/jmorton/public_www/deepblast-public-data/tmalign.tar.gz)
 
 
 See the [Malisam](http://prodata.swmed.edu/malisam/) and [Malidup](http://prodata.swmed.edu/malidup/) websites to download their datasets.
@@ -29,21 +29,21 @@ See the [Malisam](http://prodata.swmed.edu/malisam/) and [Malidup](http://prodat
 
 # Getting started
 
-We have 2 command line scripts available, namely `deepblast-train` and `deepblast-eval`.
+We have 2 command line scripts available, namely `tmvec-train` and `tmvec-eval`.
 
 ## Pretraining
 
-`deepblast-train` takes in as input a tab-delimited format of with columns
+`tmvec-train` takes in as input a tab-delimited format of with columns
 `query_seq_id | key_seq_id | tm_score1 | tm_score2 | rmsd | sequence1 | sequence2 | alignment_string`
-See an example [here](https://raw.githubusercontent.com/flatironinstitute/deepblast/master/data/tm_align_output_10k.tab) of what this looks like. At this moment, we only support parsing the output of TM-align. The parsing script can be found under
+See an example [here](https://raw.githubusercontent.com/flatironinstitute/tmvec/master/data/tm_align_output_10k.tab) of what this looks like. At this moment, we only support parsing the output of TM-align. The parsing script can be found under
 
-`deepblast/dataset/parse_tm_align.py [fname] [output_table]`
+`tmvec/dataset/parse_tm_align.py [fname] [output_table]`
 
-Once the data is configured and split appropriately, `deepblast-train` can be run.
-The command-line options are given below (see `deepblast-train --help` for more details).
+Once the data is configured and split appropriately, `tmvec-train` can be run.
+The command-line options are given below (see `tmvec-train --help` for more details).
 
 ```
-usage: deepblast-train [-h] [--gpus GPUS] [--grad-accum GRAD_ACCUM] [--grad-clip GRAD_CLIP] [--nodes NODES] [--num-workers NUM_WORKERS] [--precision PRECISION] [--backend BACKEND]
+usage: tmvec-train [-h] [--gpus GPUS] [--grad-accum GRAD_ACCUM] [--grad-clip GRAD_CLIP] [--nodes NODES] [--num-workers NUM_WORKERS] [--precision PRECISION] [--backend BACKEND]
                        [--load-from-checkpoint LOAD_FROM_CHECKPOINT] --train-pairs TRAIN_PAIRS --test-pairs TEST_PAIRS --valid-pairs VALID_PAIRS [--embedding-dim EMBEDDING_DIM]
                        [--rnn-input-dim RNN_INPUT_DIM] [--rnn-dim RNN_DIM] [--layers LAYERS] [--loss LOSS] [--learning-rate LEARNING_RATE] [--batch-size BATCH_SIZE]
                        [--multitask MULTITASK] [--finetune FINETUNE] [--mask-gaps MASK_GAPS] [--scheduler SCHEDULER] [--epochs EPOCHS]
@@ -52,12 +52,12 @@ usage: deepblast-train [-h] [--gpus GPUS] [--grad-accum GRAD_ACCUM] [--grad-clip
 
 ## Evaluation
 
-This will evaluate how much the deepblast predictions agree with the structural alignments.
-The `deepblast-train` command will automatically evaluate the heldout test set if it completes.
-However, a separate `deepblast-evaluate` command is available in case the pretraining was interrupted.  The commandline options are given below (see `deepblast-evaluate --help` for more details)
+This will evaluate how much the tmvec predictions agree with the structural alignments.
+The `tmvec-train` command will automatically evaluate the heldout test set if it completes.
+However, a separate `tmvec-evaluate` command is available in case the pretraining was interrupted.  The commandline options are given below (see `tmvec-evaluate --help` for more details)
 
 ```
-usage: deepblast-evaluate [-h] [--gpus GPUS] [--num-workers NUM_WORKERS] [--nodes NODES] [--load-from-checkpoint LOAD_FROM_CHECKPOINT] [--precision PRECISION] [--backend BACKEND]
+usage: tmvec-evaluate [-h] [--gpus GPUS] [--num-workers NUM_WORKERS] [--nodes NODES] [--load-from-checkpoint LOAD_FROM_CHECKPOINT] [--precision PRECISION] [--backend BACKEND]
                           --train-pairs TRAIN_PAIRS --test-pairs TEST_PAIRS --valid-pairs VALID_PAIRS [--embedding-dim EMBEDDING_DIM] [--rnn-input-dim RNN_INPUT_DIM]
                           [--rnn-dim RNN_DIM] [--layers LAYERS] [--loss LOSS] [--learning-rate LEARNING_RATE] [--batch-size BATCH_SIZE] [--multitask MULTITASK]
                           [--finetune FINETUNE] [--mask-gaps MASK_GAPS] [--scheduler SCHEDULER] [--epochs EPOCHS] [--visualization-fraction VISUALIZATION_FRACTION] -o
@@ -70,9 +70,9 @@ usage: deepblast-evaluate [-h] [--gpus GPUS] [--num-workers NUM_WORKERS] [--node
 ```python
 
 import torch
-from deepblast.trainer import LightningAligner
-from deepblast.dataset.utils import pack_sequences
-from deepblast.dataset.utils import states2alignment
+from tmvec.trainer import LightningAligner
+from tmvec.dataset.utils import pack_sequences
+from tmvec.dataset.utils import states2alignment
 import matplotlib.pyplot as plt
 import seaborn as sns
 
@@ -126,7 +126,7 @@ QQNKELNFKLREKQNEIFELKKIAETLRSKLEKYVDITKKLEDQNLNLQIKISDLEKKLSDA
 Score 282.3163757324219
 ```
 
-![](https://raw.githubusercontent.com/flatironinstitute/deepblast/master/imgs/example-alignment.png "example alignment")
+![](https://raw.githubusercontent.com/flatironinstitute/tmvec/master/imgs/example-alignment.png "example alignment")
 
 # FAQ
