+++ tmvec/alignment.py
@@ -1,9 +1,9 @@
 import torch
 import torch.nn as nn
-from deepblast.language_model import BiLM, pretrained_language_models
-from deepblast.nw_cuda import NeedlemanWunschDecoder as NWDecoderCUDA
-from deepblast.embedding import StackedRNN, EmbedLinear
-from deepblast.dataset.utils import unpack_sequences
+from tmvec.language_model import BiLM, pretrained_language_models
+from tmvec.nw_cuda import NeedlemanWunschDecoder as NWDecoderCUDA
+from tmvec.embedding import StackedRNN, EmbedLinear
+from tmvec.dataset.utils import unpack_sequences
 import torch.nn.functional as F
