+++ tmvec/constants.py
