+++ tmvec/utils.py
